Fix all C compiler warnings.

+++ main.c
@@ -169,7 +169,7 @@ int main(int argc, char * argv[]) {
   for(i = 2; i < argc; i++) {
     size_t fileLen = 0;
     char * fileContents = load_file(argv[i], &fileLen);
-    printf("File Length: %i chars\n", fileLen);
+    printf("File Length: %i chars\n", (int)fileLen);
     if(!gunderscript_build(&ginst, fileContents, fileLen)) {
       print_compile_error(&ginst);
       print_build_fail();
